pyproject.toml
[project]
name = "authentication-authorization-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.136.0",
    "multipart>=1.3.1",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.44.0",
]
